refactor: update credit and job service functions to use async/await
Modified functions to use async/await for credit conversion and job handling, improving code readability and ensuring proper asynchronous behavior. Updated import paths for JobWithRelations to the new types file across multiple components.
