+++ web-app/src/app/app/agents/[agentId]/jobs/@right/components/job-details.tsx
@@ -3,7 +3,7 @@ import { useFormatter, useTranslations } from "next-intl";
 import Markdown from "react-markdown";
 
 import JobStatusBadge from "@/app/agents/[agentId]/jobs/@right/components/job-status-badge";
-import { JobWithRelations } from "@/lib/db/services/job.service";
+import { JobWithRelations } from "@/lib/db/types/job.types";
 import { cn } from "@/lib/utils";
 
 interface JobDetailsProps {
