+++ web-app/src/lib/db/types/job.types.ts
@@ -1,0 +1,14 @@
+import { Prisma } from "@prisma/client";
+
+export const jobInclude = {
+  agent: true,
+  user: true,
+} as const;
+
+export const jobOrderBy = {
+  createdAt: "desc",
+} as const;
+
+export type JobWithRelations = Prisma.JobGetPayload<{
+  include: typeof jobInclude;
+}>;
